seconde tentative pour inserer et afficher des ventes mais toujours l'erreur 500

+++ dvdstore/src/main/java/com/simplon/dvdstore/dvdstore/services/DvdModelServiceMovies.java
@@ -12,6 +12,6 @@ public class DvdModelServiceMovies {
 
     String genre;
     String name;
-    int quantity;
+    Long quantity;
     float price;
 }
